move renameKey function to to utils/Obj
