+++ src/utils/Obj.ts
@@ -18,3 +18,15 @@ export const pickBy = (object, picker) => {
 
   return result;
 };
+
+const clone = (obj) => Object.assign({}, obj);
+
+export const renameKey = (object, key, newKey) => {
+  const clonedObj = clone(object);
+  const targetKey = clonedObj[key];
+
+  delete clonedObj[key];
+  clonedObj[newKey] = targetKey;
+
+  return clonedObj;
+};
